refactor to have key-value store initalized with all other adapters
