+++ busy_beaver/models/__init__.py
@@ -6,4 +6,4 @@ from .user import User  # noqa
 from .api_user import ApiUser  # noqa
 
 # Key-Value Store
-from .kv_store import kv_store  # noqa
+from .kv_store import store  # noqa
